+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add freestyle-cloudfunctions middleware type
+
 ### Changed
 - ID is now variable name friendly